Проведен небольшой рефакторинг в части загрузки изображений из формы создания опроса
