+++ backend/src/DB/db.js
@@ -1,56 +1,1 @@
-export const DB = {
-  users: [
-    {
-      name: 'Jane Doe',
-      email: 'example0@gmail.com',
-      id: 'fcc953b4-f6f8-4706-90b8-b509656ce386',
-      password: '$2b$10$JK7q3EqKdzr293tj6QXS0.WPFoaTMUiqOzAlXWmM7AdcTi.kLe1iW',
-    },
-    {
-      name: 'Jane Doe',
-      email: 'example1@gmail.com',
-      id: '09bf6c36-9c93-4ba9-b5ae-aa353e082825',
-      password: '$2b$10$qx03cntsVTmYcGxiPB4Nx.ppCrG5RCouH564FujzxgtdBQJBuGy1W',
-      accessToken:
-        'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImV4YW1wbGUxQGdtYWlsLmNvbSIsImlkIjoiMDliZjZjMzYtOWM5My00YmE5LWI1YWUtYWEzNTNlMDgyODI1IiwiaWF0IjoxNjc4NzI3NDE1LCJleHAiOjE2Nzg3MzM0MTV9.g5f6CWgzO_iT15xgP1q9uamefLvHMJ70ypHGFmxO4Iw',
-      refreshToken:
-        'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImV4YW1wbGUxQGdtYWlsLmNvbSIsImlkIjoiMDliZjZjMzYtOWM5My00YmE5LWI1YWUtYWEzNTNlMDgyODI1IiwiaWF0IjoxNjc4NzI3NDE1LCJleHAiOjE2Nzg3MzEwMTV9.tbkrz0hOPY9ZZ8mGWpbaU7GiC3SluLqVuQf1r96eH0w',
-    },
-    {
-      name: 'Jane Doe',
-      email: 'example2@gmail.com',
-      id: '8765a8f6-710e-4ae4-b583-6a37a87187e2',
-      password: '$2b$10$cG7zeyhfkwHYMDqwv5Ls3OF7m56AU0vkMIoFtpWzoIaMRLH/RevyO',
-      accessToken:
-        'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImV4YW1wbGUyQGdtYWlsLmNvbSIsImlkIjoiODc2NWE4ZjYtNzEwZS00YWU0LWI1ODMtNmEzN2E4NzE4N2UyIiwiaWF0IjoxNjc4NTM1MDc2LCJleHAiOjE2Nzg1MzUxMzZ9.6mdoyiyR_OjECzQEdKiaG8z8xipElPuQY3ILkGnsdzs',
-      refreshToken:
-        'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImV4YW1wbGUyQGdtYWlsLmNvbSIsImlkIjoiODc2NWE4ZjYtNzEwZS00YWU0LWI1ODMtNmEzN2E4NzE4N2UyIiwiaWF0IjoxNjc4NTM1MDc2LCJleHAiOjE2Nzg1Mzg2NzZ9.CAv1QMpCCMztJLpduTBR3yK1g0W1J4Wf2W3ACT7M5U0',
-    },
-    {
-      name: 'Анна',
-      email: 'abachledacurus@gmail.com',
-      id: 'f956c918-6382-4222-b84a-045fdacadc99',
-      password: '$2b$10$BMgwOUwLVo.ihkIuNMM9Q.6r3OoMDuHDVQXxdbhHoKm809nl0iHri',
-    },
-  ],
-  surveys: [
-    {
-      title: 'hrh',
-      author: '09bf6c36-9c93-4ba9-b5ae-aa353e082825',
-      surveyType: 'MC',
-      extraOption: false,
-      surveyId: 'MC952df335-aa8f-4f0c-bebc-0c951a28c9a3',
-      done: [],
-      visited: [],
-      options: [
-        {
-          optionTitle: 'hh',
-          activeLink: '',
-          image: '841fa7b56c178a3fc7f6b5788f4697eb.jpg',
-          optionId: '3ef08958-fe16-480b-b09a-7d02e648c609',
-          checked: [],
-        },
-      ],
-    },
-  ],
-};
+export const DB = {"users":[{"name":"Jane Doe","email":"example0@gmail.com","id":"fcc953b4-f6f8-4706-90b8-b509656ce386","password":"$2b$10$JK7q3EqKdzr293tj6QXS0.WPFoaTMUiqOzAlXWmM7AdcTi.kLe1iW"},{"name":"Jane Doe","email":"example1@gmail.com","id":"09bf6c36-9c93-4ba9-b5ae-aa353e082825","password":"$2b$10$qx03cntsVTmYcGxiPB4Nx.ppCrG5RCouH564FujzxgtdBQJBuGy1W","accessToken":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImV4YW1wbGUxQGdtYWlsLmNvbSIsImlkIjoiMDliZjZjMzYtOWM5My00YmE5LWI1YWUtYWEzNTNlMDgyODI1IiwiaWF0IjoxNjc4NzI3NDE1LCJleHAiOjE2Nzg3MzM0MTV9.g5f6CWgzO_iT15xgP1q9uamefLvHMJ70ypHGFmxO4Iw","refreshToken":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImV4YW1wbGUxQGdtYWlsLmNvbSIsImlkIjoiMDliZjZjMzYtOWM5My00YmE5LWI1YWUtYWEzNTNlMDgyODI1IiwiaWF0IjoxNjc4NzI3NDE1LCJleHAiOjE2Nzg3MzEwMTV9.tbkrz0hOPY9ZZ8mGWpbaU7GiC3SluLqVuQf1r96eH0w"},{"name":"Jane Doe","email":"example2@gmail.com","id":"8765a8f6-710e-4ae4-b583-6a37a87187e2","password":"$2b$10$cG7zeyhfkwHYMDqwv5Ls3OF7m56AU0vkMIoFtpWzoIaMRLH/RevyO","accessToken":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImV4YW1wbGUyQGdtYWlsLmNvbSIsImlkIjoiODc2NWE4ZjYtNzEwZS00YWU0LWI1ODMtNmEzN2E4NzE4N2UyIiwiaWF0IjoxNjc4NTM1MDc2LCJleHAiOjE2Nzg1MzUxMzZ9.6mdoyiyR_OjECzQEdKiaG8z8xipElPuQY3ILkGnsdzs","refreshToken":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImV4YW1wbGUyQGdtYWlsLmNvbSIsImlkIjoiODc2NWE4ZjYtNzEwZS00YWU0LWI1ODMtNmEzN2E4NzE4N2UyIiwiaWF0IjoxNjc4NTM1MDc2LCJleHAiOjE2Nzg1Mzg2NzZ9.CAv1QMpCCMztJLpduTBR3yK1g0W1J4Wf2W3ACT7M5U0"},{"name":"Анна","email":"abachledacurus@gmail.com","id":"f956c918-6382-4222-b84a-045fdacadc99","password":"$2b$10$BMgwOUwLVo.ihkIuNMM9Q.6r3OoMDuHDVQXxdbhHoKm809nl0iHri","accessToken":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFiYWNobGVkYWN1cnVzQGdtYWlsLmNvbSIsImlkIjoiZjk1NmM5MTgtNjM4Mi00MjIyLWI4NGEtMDQ1ZmRhY2FkYzk5IiwiaWF0IjoxNjc4NzMyMDA2LCJleHAiOjE2Nzg3MzIwNjZ9.mUNujPR_efy4o78SMgQgl0YOyp-oEXZVSGRHBTLL0QI","refreshToken":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFiYWNobGVkYWN1cnVzQGdtYWlsLmNvbSIsImlkIjoiZjk1NmM5MTgtNjM4Mi00MjIyLWI4NGEtMDQ1ZmRhY2FkYzk5IiwiaWF0IjoxNjc4NzMyMDA2LCJleHAiOjE2Nzg3MzU2MDZ9.quEGw8hNnIfH4Ub80rkANIPMteL3FKAm8jbx8vU8O9U"}],"surveys":[{"title":"hrh","author":"09bf6c36-9c93-4ba9-b5ae-aa353e082825","surveyType":"MC","extraOption":false,"surveyId":"MC952df335-aa8f-4f0c-bebc-0c951a28c9a3","done":[],"visited":[],"options":[{"optionTitle":"hh","activeLink":"","image":"841fa7b56c178a3fc7f6b5788f4697eb.jpg","optionId":"3ef08958-fe16-480b-b09a-7d02e648c609","checked":[]}]},{"title":"jjt","author":"09bf6c36-9c93-4ba9-b5ae-aa353e082825","surveyType":"MC","extraOption":false,"surveyId":"MC67651f64-d988-407c-831a-9a401757bd33","done":[],"visited":[],"options":[{"optionTitle":"jrr","activeLink":"","image":"","optionId":"a9e175bf-431d-4af4-ab5b-092f26298e3d","checked":[]}]}]}
